Setting a time trap did not always work
EUnit applies a rather short default timeout to all tests. You can
override it with a *shorter* timeout on a higher level, but if you want
a *longer* timeout you must do it directly around the actual test
representation.

I chose to work around this by wrapping each test case in a one hour
timeout. That way you can always override it with a shorter timeout.